edges partially work
