+++ packages/client/hmi-client/src/components/model-template/tera-model-templator.vue
@@ -1,0 +1,289 @@
+<template>
+	<tera-infinite-canvas
+		@click="onCanvasClick"
+		@save-transform="saveTransform"
+		@mouseenter="setMouseOverCanvas(true)"
+		@mouseleave="setMouseOverCanvas(false)"
+		@drop="onDrop"
+		@dragover.prevent
+		@dragenter.prevent
+		@focus="() => {}"
+		@blur="() => {}"
+	>
+		<template #foreground>
+			<aside>
+				<section v-if="model?.header?.schema_name">
+					<header>Model framework</header>
+					<h5>{{ model.header.schema_name }}<i class="pi pi-info-circle"></i></h5>
+				</section>
+				<section>
+					<header>Model templates</header>
+					<ul>
+						<li v-for="(_, index) in 5" :key="index">
+							<tera-model-template :card="newCard" draggable="true" />
+						</li>
+					</ul>
+				</section>
+				{{ edges }}
+				<section class="trash">
+					<i class="pi pi-trash"></i>
+					<div>Drag items here to delete</div>
+				</section>
+			</aside>
+		</template>
+		<template #data>
+			<tera-canvas-item
+				v-for="(card, index) in cards"
+				:key="index"
+				:style="{
+					width: 'fit-content',
+					top: `${card.y}px`,
+					left: `${card.x}px`
+				}"
+				@dragging="(event) => updatePosition(card, event)"
+			>
+				<tera-model-template
+					:card="card"
+					@port-selected="createNewEdge(card)"
+					@port-mouseover="onPortMouseover"
+					@port-mouseleave="onPortMouseleave"
+				/>
+			</tera-canvas-item>
+		</template>
+		<template #background>
+			<path
+				v-if="newEdge?.points"
+				:d="drawPath(interpolatePointsForCurve(newEdge.points[0], newEdge.points[1]))"
+				stroke="#1B8073"
+				stroke-width="2"
+				marker-start="url(#circle)"
+				marker-end="url(#arrow)"
+				fill="none"
+			/>
+			<path
+				v-for="(edge, index) in edges"
+				:d="drawPath(interpolatePointsForCurve(edge.points[0], edge.points[1]))"
+				stroke="#1B8073"
+				stroke-width="2"
+				marker-start="url(#circle)"
+				:key="index"
+				fill="none"
+			/>
+		</template>
+	</tera-infinite-canvas>
+</template>
+
+<script setup lang="ts">
+import { ref, computed, onMounted, onUnmounted } from 'vue';
+import * as d3 from 'd3';
+import { Position } from '@/types/workflow'; // temp
+import { Model } from '@/types/Types';
+import TeraInfiniteCanvas from '../widgets/tera-infinite-canvas.vue';
+import TeraModelTemplate from './tera-model-template.vue';
+import TeraCanvasItem from '../widgets/tera-canvas-item.vue';
+
+defineProps<{
+	model?: Model;
+}>();
+
+interface ModelTemplate {
+	// Position on canvas
+	id: number;
+	x: number;
+	y: number;
+}
+
+interface ModelTemplateEdge {
+	id: number;
+	source: number;
+	target: number;
+	points: Position[];
+}
+
+let currentPortPosition: Position = { x: 0, y: 0 };
+let isMouseOverCanvas = false;
+let canvasTransform = { x: 0, y: 0, k: 1 };
+let isMouseOverPort = false;
+
+const cards = ref<ModelTemplate[]>([{ id: 1, x: 300, y: 40 }]);
+const edges = ref<ModelTemplateEdge[]>([]);
+
+const newCard: ModelTemplate = { id: -1, x: 0, y: 0 };
+const newEdge = ref();
+const isCreatingNewEdge = computed(
+	() => newEdge.value && newEdge.value.points && newEdge.value.points.length === 2
+);
+
+function interpolatePointsForCurve(a: Position, b: Position): Position[] {
+	const controlXOffset = 50;
+	return [a, { x: a.x + controlXOffset, y: a.y }, { x: b.x - controlXOffset, y: b.y }, b];
+}
+
+const pathFn = d3
+	.line<{ x: number; y: number }>()
+	.x((d) => d.x)
+	.y((d) => d.y);
+// .curve(d3.curveBasis);
+
+// Get around typescript complaints
+const drawPath = (v: any) => pathFn(v) as string;
+
+function createNewEdge(card: ModelTemplate) {
+	if (!isCreatingNewEdge.value) {
+		newEdge.value = {
+			id: 'new edge',
+			points: [
+				{ x: currentPortPosition.x, y: currentPortPosition.y },
+				{ x: currentPortPosition.x, y: currentPortPosition.y }
+			]
+		};
+	} else {
+		edges.value.push({
+			id: edges.value.length + 1,
+			points: newEdge.value.points,
+			source: card.id,
+			target: card.id
+		});
+		cancelNewEdge();
+	}
+
+	console.log(newEdge.value);
+}
+
+function onPortMouseover(position: Position) {
+	currentPortPosition = position;
+	isMouseOverPort = true;
+}
+
+function onPortMouseleave() {
+	isMouseOverPort = false;
+}
+
+function onCanvasClick() {
+	if (isCreatingNewEdge.value) {
+		cancelNewEdge();
+	}
+}
+
+function cancelNewEdge() {
+	newEdge.value = undefined;
+}
+
+const setMouseOverCanvas = (val: boolean) => {
+	isMouseOverCanvas = val;
+};
+
+function saveTransform(newTransform: { k: number; x: number; y: number }) {
+	canvasTransform = newTransform;
+}
+
+function updateNewCardPosition(event) {
+	newCard.x = (event.offsetX - canvasTransform.x) / canvasTransform.k;
+	newCard.y = (event.offsetY - canvasTransform.y) / canvasTransform.k;
+}
+
+function onDrop(event) {
+	updateNewCardPosition(event);
+	cards.value.push({ id: cards.value.length + 1, x: newCard.x, y: newCard.y });
+}
+
+let prevX = 0;
+let prevY = 0;
+function mouseUpdate(event: MouseEvent) {
+	if (isCreatingNewEdge.value) {
+		const pointIndex = 1;
+		if (isMouseOverPort) {
+			newEdge.value!.points[pointIndex].x = currentPortPosition.x;
+			newEdge.value!.points[pointIndex].y = currentPortPosition.y;
+		} else {
+			const dx = event.x - prevX;
+			const dy = event.y - prevY;
+			newEdge.value!.points[pointIndex].x += dx / canvasTransform.k;
+			newEdge.value!.points[pointIndex].y += dy / canvasTransform.k;
+		}
+	}
+	prevX = event.x;
+	prevY = event.y;
+}
+
+const updatePosition = (card: ModelTemplate, { x, y }) => {
+	if (!isMouseOverCanvas) return;
+
+	// Update node position
+	card.x += x / canvasTransform.k;
+	card.y += y / canvasTransform.k;
+
+	// Update edge positions
+	edges.value.forEach((edge) => {
+		if (edge.source === card.id) {
+			edge.points[0].x += x / canvasTransform.k;
+			edge.points[0].y += y / canvasTransform.k;
+		}
+		if (edge.target === card.id) {
+			edge.points[edge.points.length - 1].x += x / canvasTransform.k;
+			edge.points[edge.points.length - 1].y += y / canvasTransform.k;
+		}
+	});
+};
+
+onMounted(() => {
+	document.addEventListener('mousemove', mouseUpdate);
+});
+onUnmounted(() => {
+	document.removeEventListener('mousemove', mouseUpdate);
+});
+</script>
+
+<style scoped>
+aside {
+	width: 15rem;
+	display: flex;
+	flex-direction: column;
+	height: 100%;
+	background-color: #f4f7fa;
+	border-right: 1px solid var(--surface-border-alt);
+	padding: 1rem;
+	gap: 0.5rem;
+}
+
+ul {
+	list-style: none;
+	display: flex;
+	flex-direction: column;
+	gap: 0.5rem;
+	margin-top: 0.5rem;
+}
+
+h5 {
+	display: flex;
+	align-items: center;
+	gap: 0.25rem;
+	font-weight: var(--font-weight);
+}
+
+header {
+	color: var(--text-color-subdued);
+	font-size: var(--font-caption);
+}
+
+.pi-info-circle {
+	color: var(--text-color-subdued);
+	cursor: help;
+}
+
+.trash {
+	margin-top: auto;
+	font-size: var(--font-caption);
+	color: var(--text-color-subdued);
+	border: 1px dashed #9fa9b7;
+	border-radius: var(--border-radius);
+	background-color: #eff2f5;
+	text-align: center;
+	padding: 1rem 0;
+
+	& > .pi-trash {
+		font-size: 1.5rem;
+		margin-bottom: 0.5rem;
+	}
+}
+</style>
