+++ packages/client/hmi-client/src/temp/model-template-test.vue
@@ -1,7 +1,7 @@
 <template>
-	<tera-model-template />
+	<tera-model-templator />
 </template>
 
 <script setup lang="ts">
-import TeraModelTemplate from '@/components/model-template/tera-model-template.vue';
+import TeraModelTemplator from '@/components/model-template/tera-model-templator.vue';
 </script>
